lets revert BigDecimal to Double, it was an overkill

+++ app/src/main/kotlin/com/simplemobiletools/calculator/operation/OperationFactory.kt
@@ -2,11 +2,10 @@ package com.simplemobiletools.calculator.operation
 
 import com.simplemobiletools.calculator.helpers.*
 import com.simplemobiletools.calculator.operation.base.Operation
-import java.math.BigDecimal
 
 object OperationFactory {
 
-    fun forId(id: String, baseValue: BigDecimal, secondValue: BigDecimal): Operation? {
+    fun forId(id: String, baseValue: Double, secondValue: Double): Operation? {
         return when (id) {
             PLUS -> PlusOperation(baseValue, secondValue)
             MINUS -> MinusOperation(baseValue, secondValue)
