+++ app/src/main/kotlin/com/simplemobiletools/calculator/operation/base/BinaryOperation.kt
@@ -1,5 +1,3 @@
 package com.simplemobiletools.calculator.operation.base
 
-import java.math.BigDecimal
-
-open class BinaryOperation protected constructor(protected val baseValue: BigDecimal, protected val secondValue: BigDecimal)
+open class BinaryOperation protected constructor(protected val baseValue: Double, protected val secondValue: Double)
